modified the test module
